src/app/services/characters/service.test.ts
import {describe, expect, it, jest} from "@jest/globals"
import axios from "axios"
import * as swapiClient from '@/app/api/swapiClient'
import {getCharacterById, findCharacterByName} from "@/app/services/characters/service"
import {SwapiCharacter, SwapiFilm, SwapiSpecies, SwapiStarship} from "@/app/api/swapiClient/types"

jest.mock('axios');
jest.mock('@/app/api/swapiClient');

// TODO: module mocking isn't working skipping for now
describe.skip('Character service', () => {
  const mockGetCharacterById = swapiClient.getCharacterById as jest.Mock<typeof swapiClient.getCharacterById>;
  const mockFindCharacterByName = swapiClient.findCharacterByName as jest.Mock<typeof swapiClient.findCharacterByName>;
  const mockAxiosGet = axios.get as jest.Mock<typeof axios.get>;

  it('should get character by id', async () => {
    const mockCharacter: SwapiCharacter = { url: 'https://swapi.dev/api/people/1/', name: 'Luke Skywalker', films: [], starships: [], species: [] } as unknown as SwapiCharacter
    mockGetCharacterById.mockResolvedValue(mockCharacter);
    const result = await getCharacterById('1');
    expect(result).toEqual({ id: '1', name: 'Luke Skywalker', films: [], starships: [], species: [] });
  });

  it('should find character by name', async () => {
    const mockCharacters: SwapiCharacter[] = [{ url: 'https://swapi.dev/api/people/1/', name: 'Luke Skywalker' } as unknown as SwapiCharacter]
    mockFindCharacterByName.mockResolvedValue(mockCharacters);
    const result = await findCharacterByName('Luke Skywalker');
    expect(result).toEqual([{ id: '1', name: 'Luke Skywalker' }]);
  });

  it('should handle character with films', async () => {
    const mockCharacter: SwapiCharacter = { url: 'https://swapi.dev/api/people/1/', name: 'Luke Skywalker', films: ['https://swapi.dev/api/films/1/'], starships: [], species: [] } as unknown as SwapiCharacter
    const mockFilm: SwapiFilm = { url: 'https://swapi.dev/api/films/1/', title: 'A New Hope' };
    mockGetCharacterById.mockResolvedValue(mockCharacter);
    mockAxiosGet.mockResolvedValue({ data: mockFilm });
    const result = await getCharacterById('1');
    expect(result).toEqual({ id: '1', name: 'Luke Skywalker', films: [{ id: '1', title: 'A New Hope' }], starships: [], species: [] });
  });

  it('should handle character with starships', async () => {
    const mockCharacter: SwapiCharacter = { url: 'https://swapi.dev/api/people/1/', name: 'Luke Skywalker', films: [], starships: ['https://swapi.dev/api/starships/1/'], species: [] } as unknown as SwapiCharacter
    const mockStarship: SwapiStarship = { url: 'https://swapi.dev/api/starships/1/', name: 'Millennium Falcon' };
    mockGetCharacterById.mockResolvedValue(mockCharacter);
    mockAxiosGet.mockResolvedValue({ data: mockStarship });
    const result = await getCharacterById('1');
    expect(result).toEqual({ id: '1', name: 'Luke Skywalker', films: [], starships: [{ id: '1', name: 'Millennium Falcon' }], species: [] });
  });

  it('should handle character with species', async () => {
    const mockCharacter: SwapiCharacter = { url: 'https://swapi.dev/api/people/1/', name: 'Luke Skywalker', films: [], starships: [], species: ['https://swapi.dev/api/species/1/'] } as unknown as SwapiCharacter
    const mockSpecies: SwapiSpecies = { url: 'https://swapi.dev/api/species/1/', name: 'Human' };
    mockGetCharacterById.mockResolvedValue(mockCharacter);
    mockAxiosGet.mockResolvedValue({ data: mockSpecies });
    const result = await getCharacterById('1');
    expect(result).toEqual({ id: '1', name: 'Luke Skywalker', films: [], starships: [], species: [{ id: '1', name: 'Human' }] });
  });
});
